+++ packages/react/src/components/FirstTimeOrientation/__stories__/FirstTimeOrientation.stories.jsx
@@ -46,7 +46,7 @@ export const FirstTimeOrientationStory = () => {
         Show Interstitial modal
       </Button>
       <InterstitialScreen
-        isOpen={showInterstitialModal}
+        open={showInterstitialModal}
         onClose={() => {
           setShowInterstitialModal(false);
         }}