+++ packages/react/src/components/FirstTimeOrientation/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.9.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-first-time-orientation@0.8.0...@carbon-labs/react-first-time-orientation@0.9.0) (2025-08-25)
+
+**Note:** Version bump only for package @carbon-labs/react-first-time-orientation
+
+
+
+
+
+# [0.8.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-first-time-orientation@0.7.0...@carbon-labs/react-first-time-orientation@0.8.0) (2025-08-19)
+
+
+### Bug Fixes
+
+* **firstTimeOrientation:** pattern not opening after upgrade ([#755](https://github.com/carbon-design-system/carbon-labs/issues/755)) ([758cf42](https://github.com/carbon-design-system/carbon-labs/commit/758cf4239902206b72cb9e098a94f140bdb0d4d2))
+
+
+
+
+
 # 0.7.0 (2025-07-14)